Cleaned up code to mollify MSVC++. This revision has compilation errors introduced in earlier revisions.
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@4 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: 6f8d9e70d8628542f7a576310ff93815160864e6

+++ basecode/Element.cpp
@@ -49,7 +49,7 @@ const Cinfo Element::cinfo_(
 //////////////////////////////////////////////////////////////////
 Element* lookupChildIn(const Conn *m)
 {
-	static const unsigned long OFFSET =
+	static const size_t OFFSET =
 		FIELD_OFFSET( Element, childIn_ );
 //		(unsigned long)(&Element::childIn_);
 // Need to figure out how to do this using static casts.
